Issue #9567 remove tests for non-chosen fail over approach
