+++ dev/com.ibm.ws.concurrent.persistent_fat_failover1serv/publish/servers/com.ibm.ws.concurrent.persistent.fat.failover1serv/server.xml
@@ -21,10 +21,10 @@
   <include location="../fatTestPorts.xml"/>
 
   <!-- cannot run tasks -->
-  <persistentExecutor id="persistentExec1" jndiName="persistent/exec1" enableTaskExecution="false" heartbeatInterval="15s"/>
+  <persistentExecutor id="persistentExec1" jndiName="persistent/exec1" enableTaskExecution="false"/>
 
   <!-- runs its own tasks and takes late tasks from others -->
-  <persistentExecutor id="persistentExec2" jndiName="persistent/exec2" pollInterval="2s" missedTaskThreshold="3s" heartbeatInterval="1s"/>
+  <persistentExecutor id="persistentExec2" jndiName="persistent/exec2" pollInterval="2s" missedTaskThreshold="3s"/>
 
   <!-- database for scheduled tasks -->
   <dataSource id="DefaultDataSource">
