+++ dev/com.ibm.ws.concurrent.persistent_fat_failover1serv/test-applications/failover1servApp/src/failover1serv/web/Failover1ServerTestServlet.java
@@ -86,110 +86,6 @@ public class Failover1ServerTestServlet extends FATServlet {
     }
 
     /**
-     * testHeartbeatsAreRepeatedlySent - verifies that heart beats are being sent periodically, with an increasing expiry timestamp.
-     */
-    public void testHeartbeatsAreRepeatedlySent() throws Exception {
-        // Ensure the database tables are present
-        PersistentExecutor executor = InitialContext.doLookup("persistent/exec2");
-        executor.getProperty("testHeartbeatsAreRepeatedlySent");
-
-        DataSource ds = InitialContext.doLookup("java:comp/DefaultDataSource");
-        try (Connection con = ds.getConnection()) {
-            PreparedStatement st = con.prepareStatement("SELECT EXPIRY FROM WLPPART WHERE EXECUTOR=?");
-            st.setString(1, "persistentExec2");
-            ResultSet result = st.executeQuery();
-            assertTrue(result.next());
-            long expiry1 = result.getLong(1);
-            st.close();
-            System.out.println("Found heartbeat from persistentExec2 with expiry at " + expiry1);
-
-            // Poll for the expiry to be updated by a subsequent heartbeat
-            st = con.prepareStatement("SELECT EXPIRY FROM WLPPART WHERE EXECUTOR=? AND EXPIRY>?");
-            st.setString(1, "persistentExec2");
-            st.setLong(2, expiry1);
-
-            boolean found = false;
-            for (long start = System.nanoTime();
-                    !(found = (result = st.executeQuery()).next()) && System.nanoTime() - start < TIMEOUT_NS;
-                    Thread.sleep(POLL_INTERVAL_MS))
-                ;
-
-            assertTrue(found);
-            long expiry2 = result.getLong(1);
-            System.out.println("Found heartbeat from persistentExec2 with expiry at " + expiry2);
-            st.setLong(2, expiry2);
-
-            found = false;
-            for (long start = System.nanoTime();
-                    !(found = (result = st.executeQuery()).next()) && System.nanoTime() - start < TIMEOUT_NS;
-                    Thread.sleep(POLL_INTERVAL_MS))
-                ;
-
-            assertTrue(found);
-            long expiry3 = result.getLong(1);
-            System.out.println("Found heartbeat from persistentExec2 with expiry at " + expiry3);
-        }
-    }
-
-    /**
-     * testMissedHeartbeatsClearOldPartitionData - insert entries representing missed heartbeats directly into the
-     * database. Verify that they are automatically removed (happens when heartbeat information is checked).
-     */
-    public void testMissedHeartbeatsClearOldPartitionData(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        // Ensure the database tables are present
-        PersistentExecutor executor = InitialContext.doLookup("persistent/exec2");
-        executor.getProperty("testMissedHeartbeatsClearOldPartitionData");
-
-        DataSource ds = InitialContext.doLookup("java:comp/DefaultDataSource");
-        try (Connection con = ds.getConnection()) {
-            PreparedStatement st = con.prepareStatement("INSERT INTO WLPPART(EXECUTOR,HOSTNAME,LSERVER,USERDIR,EXPIRY,STATES) VALUES(?,?,?,?,?,?)");
-
-            // insert partition data to simulate a server that stopped sending heartbeats 100 days ago
-            st.setString(1, "oldExecutor1");
-            st.setString(2, "host1.rchland.ibm.com");
-            st.setString(3, "oldServer1");
-            st.setString(4, "/Users/old1/wlp/usr");
-            st.setLong(5, System.currentTimeMillis() - TimeUnit.DAYS.toMillis(100));
-            st.setLong(6, 0);
-            assertEquals(1, st.executeUpdate());
-
-            // insert partition data to simulate a server that stopped sending heartbeats 200 days ago
-            st.setString(1, "oldExecutor2");
-            st.setString(2, "host2.rchland.ibm.com");
-            st.setString(3, "oldServer2");
-            st.setString(4, "/Users/old2/wlp/usr");
-            st.setLong(5, System.currentTimeMillis() - TimeUnit.DAYS.toMillis(200));
-            st.setLong(6, 0);
-            assertEquals(1, st.executeUpdate());
-
-            // insert partition data to simulate another server that stopped sending heartbeats 200 days ago
-            st.setString(1, "oldExecutor3");
-            st.setString(2, "host2.rchland.ibm.com");
-            st.setString(3, "oldServer3");
-            st.setString(4, "/Users/old2/wlp/usr");
-            st.setLong(5, System.currentTimeMillis() - TimeUnit.DAYS.toMillis(200));
-            st.setLong(6, 0);
-            assertEquals(1, st.executeUpdate());
-            st.close();
-
-            // Wait for the above data to be removed due to missed heartbeats
-            st = con.prepareStatement("SELECT EXECUTOR FROM WLPPART WHERE LSERVER LIKE 'oldServer%'");
-            boolean found = true;
-            for (long start = System.nanoTime(); (found = st.executeQuery().next()) && System.nanoTime() - start < TIMEOUT_NS; Thread.sleep(POLL_INTERVAL_MS))
-                ;
-
-            if (found) {
-                StringBuilder sb = new StringBuilder();
-                ResultSet result = st.executeQuery();
-                while (result.next())
-                    sb.append(result.getString(1)).append(' ');
-                result.close();
-                assertEquals("The following entries should have been removed upon detecting missed heartbeats: " + sb.toString(), 0, sb.length());
-            }
-        }
-    }
-
-    /**
      * testRemovePartition - removes the specified partition entry.
      */
     public void testRemovePartition(HttpServletRequest request, HttpServletResponse response) throws Exception {
